Add a warmup phase
In the warmup phase we only create connections and establish TLS
sessions. When the phase is over we create file events for all connected
sockets and start sending requests.

Warmup phase is ended in these cases:
1. All requested connections are connected and TLS sessions are
   established;
2. Timeout.

The main purpose of this phase is to separate CPU heavy TLS handshakes
and actual benchmarking.

This commit also removes incorrect implementation of collecting timeout
errors.